+++ packages/crawler-py/shroodler/probes/__init__.py
@@ -16,11 +16,12 @@ from shroodler.probes.mass_assignment import probe_mass_assignment
 from shroodler.probes.open_redirect import probe_open_redirect
 from shroodler.probes.path_traversal import probe_path_traversal
 from shroodler.probes.prototype_pollution import probe_prototype_pollution
-from shroodler.probes.rate_limit import probe_rate_limit
+from shroodler.probes.rate_limit import probe_rate_limit, probe_rate_limit_bypass
 from shroodler.probes.smuggling import probe_smuggling
 from shroodler.probes.sqli import probe_sqli
 from shroodler.probes.ssrf import probe_ssrf
 from shroodler.probes.ssti import probe_ssti
+from shroodler.probes.unauth_exposure import probe_unauth_exposure
 from shroodler.probes.websocket import probe_websocket
 from shroodler.probes.xss import probe_xss
 from shroodler.probes.xxe import probe_xxe
@@ -37,6 +38,8 @@ __all__ = [
     "probe_path_traversal",
     "probe_prototype_pollution",
     "probe_rate_limit",
+    "probe_rate_limit_bypass",
+    "probe_unauth_exposure",
     "probe_smuggling",
     "probe_sqli",
     "probe_ssrf",